Part 3 - An In-Memory, Append-Only, Single-Table Database
For now,
- support two operations: inserting a row and printing all rows
- reside only in memory (no persistence to disk)
- support a single, hard-coded table

insert statements are now going to look like this: insert 1 cstack
foo@bar.com

SQLite uses a B-tree for fast lookups, inserts and deletes. We’ll
start with  something simpler. Like a B-tree, it will group rows into
pages, but instead of arranging those pages as a tree it will arrange
them as an array.