Part 5 - Persistence to Disk
We make an abstraction called the pager. We ask the pager for page
number x, and the pager gives us back a block of memory. It first
looks in its cache. On a cache miss, it copies data from disk into
memory (by reading the database file).

db_open():
- opening the database file
- initializing a pager data structure
- initializing a table data structure

get_page() method has the logic for handling a cache miss

db_close():
- flushes the page cache to disk
- closes the database file
- frees the memory for the Pager and Table data structures